fix: move domains to /redocly dir, add docs, install https-proxy-agent in packages/core
