+++ packages/core/package.json
@@ -36,6 +36,7 @@
   "dependencies": {
     "@redocly/ajv": "^8.11.0",
     "colorette": "^1.2.0",
+    "https-proxy-agent": "^7.0.4",
     "js-levenshtein": "^1.1.6",
     "js-yaml": "^4.1.0",
     "lodash.isequal": "^4.5.0",
