+++ packages/core/src/decorators/common/registry-dependencies.ts
@@ -1,5 +1,5 @@
 import { UserContext } from '../../walk';
-import { isRedoclyRegistryURL } from '../../utils';
+import { isRedoclyRegistryURL } from '../../redocly/domains';
 
 import { Oas3Decorator, Oas2Decorator } from '../../visitors';
