+++ packages/core/src/redocly/__tests__/redocly-client.test.ts
@@ -1,4 +1,4 @@
-import { setRedoclyDomain } from '../../domains';
+import { setRedoclyDomain } from '../domains';
 import { RedoclyClient } from '../index';
 
 jest.mock('node-fetch', () => ({
